chore: move split components into new files
